+++ packages/ts-components/src/components/features-carousel/FeaturesCarousel.tsx
@@ -20,15 +20,17 @@ export type FeaturesCarouselItemProps = {
 type FeaturesCarouselProps = {
   heading: string;
   carouselItems: FeaturesCarouselItemProps[];
+  hideHeading: boolean;
 };
 
 export const FeaturesCarousel: React.FC<FeaturesCarouselProps> = ({
   heading,
-  carouselItems
+  carouselItems,
+  hideHeading
 }) =>
   carouselItems ? (
     <div id="benefits-carousel">
-      <Header>{heading}</Header>
+      {!hideHeading && <Header>{heading}</Header>}
       <FeaturesCarouselContainer>
         <FeaturesCarouselWindow>
           {carouselItems.map(carouselItem => (