+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.44.24](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.44.23...@times-components/article-comments@0.44.24) (2024-12-02)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 ## [0.44.23](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.44.22...@times-components/article-comments@0.44.23) (2024-11-22)
 
 **Note:** Version bump only for package @times-components/article-comments