+++ packages/article-list/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.19.26](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.25...@times-components/article-list@9.19.26) (2024-12-02)
+
+**Note:** Version bump only for package @times-components/article-list
+
+
+
+
+
 ## [9.19.25](https://github.com/newsuk/times-components/compare/@times-components/article-list@9.19.24...@times-components/article-list@9.19.25) (2024-11-22)
 
 **Note:** Version bump only for package @times-components/article-list